+++ wallet-console/src/main/java/com/isa/jjdzr/brokerlogic/BrokerLogicBuy.java
@@ -1,4 +1,4 @@
-package com.isa.jjdzr.brokerLogic;
+package com.isa.jjdzr.brokerlogic;
 
 import com.isa.jjdzr.dto.Asset;
 import com.isa.jjdzr.market.Market;
