+++ wallet-console/src/main/java/com/isa/jjdzr/consolepresentation/WalletGenView.java
@@ -1,11 +1,12 @@
-package com.isa.jjdzr.consolePresentation;
+package com.isa.jjdzr.consolepresentation;
 
 import com.isa.jjdzr.dto.Wallet;
+import com.isa.jjdzr.walletgenerator.WalletGenerator;
 
 import java.util.Scanner;
 import java.util.regex.Pattern;
 
-public class WalletGenView {
+class WalletGenView {
     public Wallet start(Wallet wallet) {
         System.out.println("Witam w generatorze portfela inwestycyjnego.");
         if (doYouWantToContinue()) {
@@ -18,7 +19,9 @@ public class WalletGenView {
                     return wallet;
                 }
             }
-            wallet = new WalletGenerator().generateWallet();
+            System.out.println("Utworzymy teraz nowy portfel inwestycyjny.");
+            String cash = getCashAmount();
+            wallet = new WalletGenerator().generateWallet(cash);
         }
         return wallet;
     }
@@ -43,6 +46,42 @@ public class WalletGenView {
         System.out.println("Portfel nie zostanie utworzony.");
         System.out.println("Powrót do menu głównego");
     }
+    private String getCashAmount() {
+        Scanner scan = new Scanner(System.in);
+        String cash;
+        System.out.println("Podaj jaką kwotą chcesz zasilić portfel: ");
+        cash = scan.nextLine();
+        cash = replaceComma(cash);
+        while (isInvalidCash(cash)) {
+            System.err.println("Nieprawidłowa wartość. Podaj kwotę: ");
+            cash = scan.nextLine();
+            cash = replaceComma(cash);
+        }
+        System.out.println("Portfel zasilony kwotą: " + cash);
+        System.out.println("Powrót do menu głównego.");
 
+        return cash;
+    }
+
+    private boolean isInvalidCash(String str) {
+        int countDots = (int) str.chars().filter(ch -> ch == '.').count();
+        if (countDots > 1) {
+            return true;
+        } else if (str.length() > 20) {
+            System.err.println("Zbyt duża kwota!");
+            return true;
+        } else if (str.charAt(0) == ('0')) {
+            System.err.println("Kwota nie może zaczynać się od 0!");
+            return true;
+        } else {
+            String validSymbols = "[0-9.]*";
+            return !Pattern.matches(validSymbols, str);
+        }
+    }
+
+    private String replaceComma(String str){
+        return str.replace(',','.');
+    }
+}
 
 }
